+++ client/app/form-builder/form-builder-sidenav.controller.js
@@ -49,7 +49,7 @@
         multipleAnswers:    'selectedNode.lfData.basic.itemHash["/multipleAnswers/1"].value',
         defaultAnswer:      'selectedNode.lfData.basic.itemHash["/defaultAnswer/1"].value',
         units:              'selectedNode.lfData.basic.itemHash["/units/1"].value',
-        calculationMethod:  'selectedNode.lfData.basic.itemHash["/calculationMethod/1"].value',
+        _calculationMethod: 'selectedNode.lfData.basic.itemHash["/_calculationMethod/1"].value',
 
         // Advanced tab
         useRestrictions:    'selectedNode.lfData.advanced.itemHash["/useRestrictions/1"].value',
@@ -138,7 +138,10 @@
             $scope.changeThisAndAncestralCustomCodes($scope.selectedNode);
             $scope.selectedNode.isDirty = false;
           }
-          formBuilderService.processNodeTree($scope.formBuilderData.treeData[0].nodes);
+          if($scope.selectedNode.skipLogicDirty) {
+            formBuilderService.processNodeTree($scope.formBuilderData.treeData[0].nodes, $scope.selectedNode);
+            $scope.selectedNode.skipLogicDirty = false;
+          }
         }
 
         // Select new node
@@ -403,7 +406,7 @@
                 return;
               }
               try {
-                formBuilderService.adjustFieldsToImportedLoinc(response);
+                formBuilderService.adjustFieldsInImportedLoinc(response);
                 updateFormBuilder(scope, response);
               }
               finally {
@@ -724,6 +727,7 @@
                   scope.selectedNode.lfData.advanced.name = scope.selectedNode.lfData.basic.name;
                   if(newValue !== oldValue) {
                     scope.selectedNode.isDirty = true;
+                    scope.selectedNode.skipLogicDirty = true;
                   }
                 }
               }, true);
@@ -739,6 +743,10 @@
                   scope.selectedNode.lfData.basic.name = scope.selectedNode.id + ' ' +
                     scope.selectedNode.lfData.basic.itemHash['/question/1'].value + code;
                   scope.selectedNode.lfData.advanced.name = scope.selectedNode.lfData.basic.name;
+                  if(newValue !== oldValue) {
+                    scope.selectedNode.isDirty = true;
+                    scope.selectedNode.skipLogicDirty = true;
+                  }
                 }
               }, true);
               break;
@@ -751,6 +759,7 @@
                     if(newValue.code === dataConstants.CUSTOM) {
                       scope.selectedNode.lfData.basic.itemHash['/questionCode/1'].editable = '1';
                       scope.selectedNode.isDirty = true;
+                      scope.selectedNode.skipLogicDirty = true;
                     }
                     else if(newValue.code === dataConstants.LOINC) {
                       scope.selectedNode.lfData.basic.itemHash['/questionCode/1'].editable = '0';
@@ -770,6 +779,7 @@
                   if (isDirty(oldValue, newValue)) {
                     var dataType = scope.selectedNode.lfData.basic.itemHash['/dataType/1'];
                     scope.selectedNode.isDirty = true;
+                    scope.selectedNode.skipLogicDirty = true;
                   }
                 }
               }, true);
@@ -791,6 +801,7 @@
 
                   if(isDirty(oldValue, newValue)) {
                     scope.selectedNode.isDirty = true;
+                    scope.selectedNode.skipLogicDirty = true;
                   }
                 }
               }, true);